backend/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
Flask-Jsonpify==1.5.0
langchain-core==0.3.47
langgraph==0.3.18
mysql-connector-python==9.2.0
langchain-google-genai==2.1.1
google-ai-generativelanguage==0.6.17
flask-cors==5.0.1
pillow==11.1.0